requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
distlib==0.3.1
Django==3.1.5
dlib==19.22.0
dnspython==1.16.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
pymongo==3.11.3
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
Werkzeug==1.0.1
